pyproject.toml
[project]
name = "newspicker"
version = "0.1.0"
description = "AI news podcast generator — collect, summarise, synthesise, and publish daily episodes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "requests>=2.32",
  "feedparser>=6.0",
  "python-dateutil>=2.9",
  "google-genai>=1.0",
  "pydub>=0.25",
  "PyYAML>=6.0",
]
